Edicion de tipos de habitacion y su galeria
Se implemento la edicion de la galeria en Google Drive de un tipo de habitación, así mismo de sus datos.

+++ Hotel-UDEC/dominionegocio/parametricaservice/src/main/java/co/edu/ucundinamarca/negocio/parametricaservice/repository/TipoHabitacionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoHabitacionRepository extends JpaRepository<TipoHabitacion, Long > {
+public interface TipoHabitacionRepository extends JpaRepository<TipoHabitacion, Integer > {
 }
